refactor util (random()
