+++ code/system/sys.util/src/Value/Value.mts
@@ -1,8 +1,10 @@
 import * as Array from './Value.Array.mjs';
 import * as Math from './Value.Math.mjs';
-import * as object from './Value.Object.mjs';
+import * as Random from './Value.Random.mjs';
 import * as To from './Value.To.mjs';
 import * as Util from './Value.Util.mjs';
+
+import * as object from './Value.Object.mjs';
 
 /**
  * Value conversion and interpretation helpers.
@@ -12,5 +14,6 @@ export const Value = {
   ...To,
   ...Array,
   ...Math,
+  ...Random,
   ...Util,
 };
